Explicitly fail when unable to obtain API user key
Without this change, if we failed to get the API user key, the next
post request that creates the paste still succeeds, but it creates
a guest paste, not associated with the user's account. This is
probably not the right thing. So this change makes the script fail
hard when it isn't able to get the API user key.